Converted things.txt to things.xml
Changed the storage file for the description of things to use XML rather
than a simple txt file. However, although the application reads in the
XML file's contents, there are only two objects and they are neither
mentioned or used in the game. The next step is to implement take,
examine, and drop.